+++ static/js/ajax/del.js
@@ -1,4 +1,0 @@
-function delete_account() {
-    // TODO: #49 delete account here
-    console.log("ajax here to server and request account deletion")
-}